LPS-197482 Implement Object Relationship Edge Component for Code Reusability
