+++ modules/apps/object/object-web/src/main/resources/META-INF/resources/js/components/ModelBuilder/Edges/DefaultObjectRelationshipEdge.tsx
@@ -3,104 +3,36 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import React, {useEffect, useMemo, useState} from 'react';
+import React, {useMemo} from 'react';
 import {
 	EdgeProps,
-	EdgeText,
+	Node,
 	getEdgeCenter,
 	getSmoothStepPath,
 	useStoreState,
 } from 'react-flow-renderer';
 
-import {useObjectFolderContext} from '../ModelBuilderContext/objectFolderContext';
-import {TYPES} from '../ModelBuilderContext/typesEnum';
 import {ObjectRelationshipEdgeData} from '../types';
 import {getEdgeParams} from '../utils';
-import ManyMarker from './ManyMarker';
-import OneMarker from './OneMarker';
-
-const DEFAULT_COLOR = '#80ACFF';
-const HIGHLIGHT_COLOR = '#0B5FFF';
-
-export function getInitialObjectRelationshipEdgeStyle(edgeSelected: boolean) {
-	return {
-		stroke: edgeSelected ? HIGHLIGHT_COLOR : DEFAULT_COLOR,
-		strokeWidth: '2px',
-	};
-}
-
-export function getInitialLabelBgStyle(edgeSelected: boolean) {
-	return {
-		fill: edgeSelected ? HIGHLIGHT_COLOR : DEFAULT_COLOR,
-		height: '24px',
-	};
-}
+import ObjectRelationshipEdge from './ObjectRelationshipEdge';
 
 export default function DefaultObjectRelationshipEdge({
 	data,
-	id,
+	id: edgeId,
 	source,
-	style = {},
 	target,
-}: EdgeProps<ObjectRelationshipEdgeData>) {
-	const {
-		label,
-		markerEndId,
-		markerStartId,
-		objectRelationshipId,
-		selected,
-	} = data!;
-
-	const [_, dispatch] = useObjectFolderContext();
-	const [
-		objectRelationshipEdgeStyle,
-		setObjectRelationshipEdgeStyle,
-	] = useState({
-		...style,
-		...getInitialObjectRelationshipEdgeStyle(selected),
-	});
-	const [labelBgStyle, setLabelBgStyle] = useState(
-		getInitialLabelBgStyle(selected)
-	);
+}: EdgeProps<ObjectRelationshipEdgeData[]>) {
 	const {nodes} = useStoreState((state) => state);
 
 	const sourceNode = useMemo(() => nodes.find((node) => node.id === source), [
 		source,
 		nodes,
-	]);
+	]) as Node<ObjectDefinitionNodeData>;
+
 	const targetNode = useMemo(() => nodes.find((node) => node.id === target), [
 		target,
 		nodes,
-	]);
-
-	useEffect(() => {
-		if (selected) {
-			setObjectRelationshipEdgeStyle((style) => {
-				return {...style, stroke: HIGHLIGHT_COLOR};
-			});
-			setLabelBgStyle((style) => {
-				return {
-					...style,
-					fill: HIGHLIGHT_COLOR,
-				};
-			});
-		}
-		else {
-			setObjectRelationshipEdgeStyle((style) => {
-				return {...style, stroke: DEFAULT_COLOR};
-			});
-			setLabelBgStyle((style) => {
-				return {
-					...style,
-					fill: DEFAULT_COLOR,
-				};
-			});
-		}
-	}, [selected]);
-
-	if (!sourceNode || !targetNode) {
-		return null;
-	}
+	]) as Node<ObjectDefinitionNodeData>;
 
 	const {
 		sourcePos,
@@ -137,49 +69,17 @@ export default function DefaultObjectRelationshipEdge({
 	});
 
 	return (
-		<g className="react-flow__connection">
-			<OneMarker />
-
-			<ManyMarker />
-
-			<path
-				className="react-flow__edge-path"
-				d={edgePath}
-				id={id}
-				markerEnd={`url(#${markerEndId})`}
-				style={objectRelationshipEdgeStyle}
-			/>
-
-			<path
-				className="react-flow__edge-path"
-				d={reverseEdgePath}
-				id={id + 'reverse'}
-				markerEnd={`url(#${markerStartId})`}
-				style={objectRelationshipEdgeStyle}
-			/>
-
-			<EdgeText
-				label={label}
-				labelBgBorderRadius={4}
-				labelBgPadding={[8, 5]}
-				labelBgStyle={labelBgStyle}
-				labelShowBg
-				labelStyle={{
-					fill: '#FFF',
-					fontSize: '12px',
-					fontWeight: 600,
-				}}
-				onClick={() => {
-					dispatch({
-						payload: {
-							selectedObjectRelationshipId: objectRelationshipId,
-						},
-						type: TYPES.SET_SELECTED_OBJECT_RELATIONSHIP_EDGE,
-					});
-				}}
-				x={edgeCenterX}
-				y={edgeCenterY}
-			/>
-		</g>
+		<ObjectRelationshipEdge
+			data={data}
+			edgeCenterX={edgeCenterX}
+			edgeCenterY={edgeCenterY}
+			edgePath={edgePath}
+			id={edgeId}
+			reverseEdgePath={reverseEdgePath}
+			sourceX={sourceX}
+			sourceY={sourceY}
+			targetX={targetX}
+			targetY={targetY}
+		/>
 	);
 }
