+++ modules/apps/object/object-web/types/src/main/resources/META-INF/resources/js/components/ModelBuilder/Edges/ObjectRelationshipEdge.d.ts
@@ -1,0 +1,27 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2024 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+/// <reference types="react" />
+
+import {EdgeProps} from 'react-flow-renderer';
+import {ObjectRelationshipEdgeData} from '../types';
+import './Edge.scss';
+interface ObjectRelationshipEdge
+	extends Partial<EdgeProps<ObjectRelationshipEdgeData[]>> {
+	edgeCenterX: number;
+	edgeCenterY: number;
+	edgePath: string;
+	reverseEdgePath?: string;
+}
+export default function ObjectRelationshipEdge({
+	edgePath,
+	edgeCenterX,
+	edgeCenterY,
+	data,
+	id: edgeId,
+	reverseEdgePath,
+	style,
+}: ObjectRelationshipEdge): JSX.Element;
+export {};
