+++ modules/apps/object/object-web/types/src/main/resources/META-INF/resources/js/components/ModelBuilder/Edges/OneMarker.d.ts
@@ -6,4 +6,10 @@
 /// <reference types="react" />
 
 export declare const oneMarkerId = 'oneMarker';
-export default function OneMarker(): JSX.Element;
+interface OneMarkerProps {
+	objectRelationshipId: string;
+}
+export default function OneMarker({
+	objectRelationshipId,
+}: OneMarkerProps): JSX.Element;
+export {};
