+++ modules/apps/object/object-web/types/src/main/resources/META-INF/resources/js/components/ModelBuilder/Edges/SelfObjectRelationshipEdge.d.ts
@@ -7,14 +7,11 @@
 
 import {EdgeProps} from 'react-flow-renderer';
 import {ObjectRelationshipEdgeData} from '../types';
-import './Edge.scss';
 export default function SelfObjectRelationshipEdge({
 	data,
-	id,
-	source,
+	id: edgeId,
 	sourceX,
 	sourceY,
-	style,
 	targetX,
 	targetY,
-}: EdgeProps<ObjectRelationshipEdgeData>): JSX.Element | null;
+}: EdgeProps<ObjectRelationshipEdgeData[]>): JSX.Element;
